feat(build): gate api/tls/http protocols behind Cargo features
Second slice of the compile-time customization work. Removes the last
"always-on" dependencies (hyper / rustls / h2) so a minimal build is a
bare DNS forwarder.

- Lift AppController + LogBuffer out of src/api/ into src/core/ so the
  core runtime (reload, shutdown, log ring buffer) no longer depends on
  the api module. src/api/{control,logs}.rs become thin HTTP adapters.
- Gate `pub mod api` behind the `api` feature; move register_plugin_api!
  / register_api_route! into src/api_macros.rs where they expand to
  Ok(()) when api is off, keeping plugin call sites cfg-free.
- Add public features api/webui/metrics, server-dot/doh,
  upstream-dot/doh, and private _tls-client/_tls-server/_http-server/
  _http-client aggregators; mark hyper/rustls/h2/etc. optional.
- Cfg-gate the matching code paths (tcp.rs DoT branch, server/http,
  conn_tcp DoT, conn_h2, UpstreamBuilder arms, http_client,
  metrics_collector) with clear "not compiled in; rebuild with
  --features ..." errors instead of panics.
- Split tls_config into client/server submodules so a TLS-server-only
  build does not pull the client webpki-roots, and a TLS-client-only
  build does not pull the server cert loaders.
- Move build_quic_endpoint into a shared server::quic_endpoint module so
  a server-doh3-only build (no server-doq) still resolves it; have
  server-doh pull dep:h2 and _http-client pull dep:tower-service so each
  protocol feature builds in isolation.
- standard bundle gains api/metrics/server-dot/doh/upstream-dot/doh;
  full adds webui. Per-feature gate the integration tests that reference
  optional plugins.

minimal release binary drops to ~8.9 MB (vs ~21 MB full, ~58% smaller).
Verified: clippy --all-targets clean for minimal/standard/all-features,
every protocol feature compiles in isolation, cargo test green across
bundles, dig works on a minimal UDP forwarder, and disabled-protocol
configs fail with clear rebuild hints. Roadmap Phase 1 marked complete.
