+++ docs/docs/custom-build.md
@@ -16,34 +16,51 @@ fork 仓库后修改 `Cargo.toml` 的 `default = [...]`,或在编译时用 `--fe
 
 | Bundle | 适用场景 | 大致内容 |
 |---|---|---|
-| `minimal` | 嵌入式 / 容器 / 学习 | UDP + TCP 监听,UDP + TCP upstream,sequence / forward / cache / fallback / hosts / redirect / arbitrary / dual_selector / ecs_handler / ttl / drop_resp / black_hole / debug_print / reload 等基础执行器,全部 matcher,`domain_set` + `ip_set` provider |
-| `standard` | 家用路由器 / 中等规模 | minimal + DoQ 上下行 + provider-protobuf(geoip/geosite/v2ray_dat) + adguard_rule + cron + script + download + http_request + reverse_lookup |
-| `full`(默认) | 全功能 | standard + DoH3 上下行 + MikroTik 集成 + query_recorder + ipset / nftset + upgrade 子命令 |
+| `minimal` | 嵌入式 / 容器 / 学习 | UDP + TCP 监听,UDP + TCP upstream,sequence / forward / cache / fallback / hosts / redirect / arbitrary / dual_selector / ecs_handler / ttl / drop_resp / black_hole / debug_print / reload 等基础执行器,全部 matcher,`domain_set` + `ip_set` provider。**不含** hyper / rustls / quinn,二进制最小 |
+| `standard` | 家用路由器 / 中等规模 | minimal + 管理 API + metrics + DoT/DoH/DoQ 上下行 + provider-protobuf(geoip/geosite/v2ray_dat) + adguard_rule + cron + script + download + http_request + reverse_lookup |
+| `full`(默认) | 全功能 | standard + WebUI + DoH3 上下行 + MikroTik 集成 + query_recorder + ipset / nftset + upgrade 子命令 |
+
+> 实测 release 二进制体积(macOS arm64,仅供参考):`minimal` ≈ 8.9 MB,
+> `standard` ≈ 17 MB,`full` ≈ 21 MB。`minimal` 把 hyper / rustls / quinn /
+> h2 / h3 / sqlite 全部排除,体积约为 `full` 的 **40%**。
 
 ## 颗粒度开关
 
 下表里的每个 feature 都可以单独打开或关闭。组合包就是这些开关的集合,
 你也可以**只挑自己需要的开关**而不走预设。
 
-### 协议
+### 入站 / 出站协议
 
 | Feature | 作用 |
 |---|---|
+| `server-dot` | 启用 DoT(TLS over TCP)入站服务器,依赖 rustls 服务端栈 |
+| `server-doh` | 启用 DoH(HTTP/2 over TLS)入站服务器,依赖 hyper 服务端 + rustls |
 | `server-doq` | 启用 DoQ(QUIC)入站服务器,依赖 `quinn` |
-| `server-doh3` | 在 DoH 服务器上启用 HTTP/3 路径,依赖 `h3` / `h3-quinn` / `quinn` |
+| `server-doh3` | 在 DoH 服务器上启用 HTTP/3 路径(需 `server-doh`),额外依赖 `h3` / `h3-quinn` / `quinn` |
+| `upstream-dot` | 启用 DoT upstream(`tls://` scheme),依赖 rustls 客户端栈 |
+| `upstream-doh` | 启用 DoH(HTTP/2)upstream(`https://` scheme),依赖 hyper-rustls + `h2` |
 | `upstream-doq` | 启用 DoQ upstream(`quic://` / `doq://` scheme) |
-| `upstream-doh3` | 启用 DoH HTTP/3 upstream(`h3://` scheme 或 `enable_http3: true`) |
+| `upstream-doh3` | 启用 DoH HTTP/3 upstream(`h3://` scheme 或 `enable_http3: true`,需 `upstream-doh`) |
 
-> 关闭 upstream-doq 后,如果 yaml 里仍写 `quic://...`,启动时会以
-> "upstream DoQ is not compiled into this build; rebuild with --features
-> upstream-doq" 报错,而不是崩溃。
+> 关闭某个协议后,如果 yaml 里仍写了对应 scheme/配置,启动时会以例如
+> "upstream DoT is not compiled into this build; rebuild with --features
+> upstream-dot" 报错,而不是崩溃。`server-dot` 关闭时,`tcp_server` 配置里
+> 写 `cert` / `key` 会得到 "DoT is not compiled into this build; rebuild with
+> --features server-dot" 的清晰提示。
 
-### 管理面 / 协议(Phase 1B)
+### 管理面
 
-`api`、`webui`、`metrics`、`server-dot`、`server-doh`、`upstream-dot`、
-`upstream-doh` 这几项需要先把 `AppController` / `LogBuffer` 从 `src/api/`
-抽到 `src/core/`,会在后续版本一并开放。本阶段它们对应的功能保持
-**默认启用且不可裁剪**。
+| Feature | 作用 | 依赖 |
+|---|---|---|
+| `api` | 管理 / 健康 / 控制 / 日志 / 配置 HTTP API,以及各插件的 `/plugins/<tag>/...` 端点 | hyper 服务端 + rustls 服务端(支持 HTTPS) |
+| `webui` | 在 API hub 上托管 WebUI 静态资源(依赖 `api`) | — |
+| `metrics` | `/metrics` Prometheus 端点 + `metrics_collector` 执行器(依赖 `api`) | — |
+
+> `api` 关闭后,`src/api/` 整个模块都不编译,hyper / rustls 服务端栈随之
+> 排除,这是 `minimal` 体积大幅下降的主因。进程内的 `MetricSource` 计数器
+> 始终保留在 core,关闭 `metrics` 只是不暴露 HTTP 端点,不影响热路径。
+> `AppController` / `LogBuffer` 已经下沉到 `src/core/`,因此核心运行时(重载、
+> 关停、日志环形缓冲)在没有 `api` 的 `minimal` 构建里依然可用。
 
 ### 可选插件
 
@@ -70,11 +87,14 @@ cargo build --release
 # 最小可用,只跑基础转发
 cargo build --release --no-default-features --features minimal
 
-# 家用路由器(含 DoQ、geo、adguard 等常用 provider/executor)
+# 家用路由器(含 API、DoT/DoH/DoQ、geo、adguard 等常用 provider/executor)
 cargo build --release --no-default-features --features standard
 
 # 只在 minimal 上加 MikroTik 集成
 cargo build --release --no-default-features --features "minimal,plugin-mikrotik"
+
+# 只要纯转发 + 管理 API,不要任何重型插件
+cargo build --release --no-default-features --features "minimal,api"
 ```
 
 ## 验证编译矩阵
