+++ src/network/mod.rs
@@ -28,8 +28,10 @@
 //! - [`upstream`]: outbound DNS resolver construction, bootstrap resolution,
 //!   and connection pooling across supported upstream protocols.
 pub mod buffer_pool;
+#[cfg(feature = "_http-client")]
 pub mod http_client;
 pub mod listen;
+#[cfg(any(feature = "_tls-client", feature = "_tls-server"))]
 pub mod tls_config;
 pub mod transport;
 pub mod upstream;
